chore(log): move createTransport into own module
